requirements.txt
jupyter
pandas
numpy
networkx
matplotlib
SQLAlchemy
pymysql
pytest
Faker
scipy
tenacity
openai
python-dotenv
pydantic
gensim
nltk